Modify RulesCheckReceiver to stage an uninstall
RulesCheckReceiver now stages an uninstall if the device
reverts to the /system version of the TimeZoneData app
rather than attempting to stage an install of the /system
version.

This reduces the number of states a "clean" device
could be in from 2 (no distro installed / system version
distro installed) to 1 (no distro installed).

It also means that in the event that a device ships with
an old or incompatible distro file in the /system version
of the .apk, an install won't even be tried (it would be
reject by the TimeZoneDistroInstaller anyway, but more
protection means libcore engineers can sleep more soundly).

Bug: 31008728
Test: Ran pending PTS test (in internal branch)
Change-Id: I1c8ecc99fb861473831352a55a400e389e98770c